+++ src/applications/CustomMessageSettingsApplication.ts
@@ -1,7 +1,11 @@
 import { weatherDescriptions } from '@/weather/weatherMap';
-import { moduleSettings, SettingKeys } from '@/settings/ModuleSettings';
+import { moduleSettings, ModuleSettingKeys } from '@/settings/ModuleSettings';
 import moduleJson from '@module';
 import { localize } from '@/utils/game';
+
+const NUM_CLIMATES = 3;
+const NUM_HUMIDITY = 3;
+const NUM_WEATHER = 37;
 
 export class CustomMessageSettingsApplication extends FormApplication {
   private _flattenedDescriptions = [] as {
@@ -59,7 +63,7 @@ export class CustomMessageSettingsApplication extends FormApplication {
 
   public async getData(): Promise<any> {
     // load any current values
-    const currentText = moduleSettings.get(SettingKeys.customChatMessages);
+    const currentText = moduleSettings.get(ModuleSettingKeys.customChatMessages);
 
     for (let i=0; i<this._flattenedDescriptions.length; i++) {
       this._flattenedDescriptions[i].currentText = currentText[this._flattenedDescriptions[i].climateId][this._flattenedDescriptions[i].humidityId][this._flattenedDescriptions[i].descriptionId];
@@ -72,16 +76,16 @@ export class CustomMessageSettingsApplication extends FormApplication {
 
   public async _updateObject(_event, formData: Record<string, string>): Promise<void> {
     // bundle up into settings object
-    const chatMessages: string[][][] = new Array(Object.keys(Climate).length/2)
+    const chatMessages: string[][][] = new Array(NUM_CLIMATES)
       .fill('')
-      .map(() => new Array(Object.keys(Humidity).length/2)
+      .map(() => new Array(NUM_HUMIDITY)
         .fill('')
-        .map(() => new Array(37).fill('')));
+        .map(() => new Array(NUM_WEATHER).fill('')));
 
     for (let i=0; i<this._flattenedDescriptions.length; i++) {
       chatMessages[this._flattenedDescriptions[i].climateId][this._flattenedDescriptions[i].humidityId][this._flattenedDescriptions[i].descriptionId] = formData['v-'+i];
     }
 
-    await moduleSettings.set(SettingKeys.customChatMessages, chatMessages);
+    await moduleSettings.set(ModuleSettingKeys.customChatMessages, chatMessages);
   }
 }